Move files for data sources in `source` directory
