+++ cli/fetch.js
@@ -14,15 +14,15 @@ var ZebraAuth = require('../lib/auth/zebra_auth');
 var GithubAuth = require('../lib/auth/github_auth');
 var GitlabAuth = require('../lib/auth/gitlab_auth');
 var NullAuth = require('../lib/auth/null_auth');
-var GoogleCalendar = require('../lib/calendar');
-var GoogleMail = require('../lib/mail');
-var Slack = require('../lib/slack');
-var Logbot = require('../lib/logbot');
-var Jira = require('../lib/jira');
-var Zebra = require('../lib/zebra');
-var Git = require('../lib/git');
-var Github = require('../lib/github');
-var Gitlab = require('../lib/gitlab');
+var GoogleCalendar = require('../lib/source/calendar');
+var GoogleMail = require('../lib/source/mail');
+var Slack = require('../lib/source/slack');
+var Logbot = require('../lib/source/logbot');
+var Jira = require('../lib/source/jira');
+var Zebra = require('../lib/source/zebra');
+var Git = require('../lib/source/git');
+var Github = require('../lib/source/github');
+var Gitlab = require('../lib/source/gitlab');
 var date = require('../lib/date');
 var util = require('../lib/util');
