+++ lib/source/calendar.js
@@ -4,7 +4,7 @@ const {google} = require('googleapis');
 var Moment = require('moment-timezone');
 var _ = require('lodash');
 
-var helper = require('./helper');
+var helper = require('../helper');
 var Source = require('./source');
 
 class GoogleCalendar extends Source {
