+++ lib/source/git.js
@@ -9,7 +9,7 @@ var Async = require('promise-async');
 var Promise = require('bluebird');
 var _ = require('lodash');
 
-var helper = require('./helper');
+var helper = require('../helper');
 var Source = require('./source');
 
 class Git extends Source {
