+++ lib/source/github.js
@@ -5,7 +5,7 @@ var Moment = require('moment-timezone');
 var Promise = require('bluebird');
 var _ = require('lodash');
 
-var helper = require('./helper');
+var helper = require('../helper');
 var Source = require('./source');
