+++ lib/source/gitlab.js
@@ -6,7 +6,7 @@ var Async = require('promise-async');
 var Promise = require('bluebird');
 var _ = require('lodash');
 
-var helper = require('./helper');
+var helper = require('../helper');
 var Source = require('./source');
 
 class Gitlab extends Source {
