+++ lib/source/jira.js
@@ -5,7 +5,7 @@ var Moment = require('moment-timezone');
 var Async = require('promise-async');
 var _ = require('lodash');
 
-var helper = require('./helper');
+var helper = require('../helper');
 var Source = require('./source');
 
 class Jira extends Source {
