+++ lib/source/logbot.js
