+++ lib/source/slack.js
