+++ lib/source/zebra.js
